Project/refactoring (#32)
* refactored MySQL db

* many2many tables have simple primary key

* applied mysql changes to JPA

* added toString to model classes

* untested: rest api changes + documentation

* refactored REST API #2

* OpenAPI documentation expansion

* UserResource documented (OpenAPI)

* Kendo Grid REST API changes + email validation

* fixed groups not getting stored upon user creation

* Note that ordered configs don't chain.

* Spring security sanity check.

* reverted to old security settings

* extended security configuration to api

* removed unnecessary exceptions

* fixed date formats

* fixed uncaught exception

* thymeleaf replace fixed

* rabbitmq set up

* rabbitMQ initial setup

* first email setup

* whitespace formatting fix

* fixed immutable groups

* refactored email

* refactored email #2

* added random url controller

* registration confirmation and password reset done

+++ src/main/java/com/samuraiDigital/adminsystem/data/repositories/UserInfoRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.samuraiDigital.adminsystem.data.model.UserInfo;
 
-public interface UserInfoRepository extends CrudRepository<UserInfo, Long>{
+public interface UserInfoRepository extends CrudRepository<UserInfo, Integer> {
 
 	Optional<UserInfo> findByName(String name);
-	
+
 }
