+++ src/main/java/com/samuraiDigital/adminsystem/data/repositories/UserSecurityDetailsRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.samuraiDigital.adminsystem.data.model.UserSecurityDetails;
 
-public interface UserSecurityDetailsRepository extends CrudRepository<UserSecurityDetails, Long>{
-	
+public interface UserSecurityDetailsRepository extends CrudRepository<UserSecurityDetails, Integer> {
+
 	Optional<UserSecurityDetails> findByUsername(String username);
 
 	Optional<UserSecurityDetails> findByEmail(String email);
-	
+
 }
